# Sliding window counter