--- a/limits/aio/storage/memory.py
+++ b/limits/aio/storage/memory.py
@@ -1,6 +1,6 @@
 import asyncio
 import time
-from collections import Counter
+from collections import Counter, defaultdict
 from math import floor
 
 from deprecated.sphinx import versionadded
@@ -42,6 +42,7 @@ class MemoryStorage(
         self, uri: Optional[str] = None, wrap_exceptions: bool = False, **_: str
     ) -> None:
         self.storage: limits.typing.Counter[str] = Counter()
+        self.locks: defaultdict[str, asyncio.Lock] = defaultdict(asyncio.Lock)
         self.expirations: Dict[str, float] = {}
         self.events: Dict[str, List[LockableEntry]] = {}
         self.timer: Optional[asyncio.Task[None]] = None
@@ -64,6 +65,7 @@ class MemoryStorage(
             if self.expirations[key] <= time.time():
                 self.storage.pop(key, None)
                 self.expirations.pop(key, None)
+                self.locks.pop(key, None)
 
     async def __schedule_expiry(self) -> None:
         if not self.timer or self.timer.done():
@@ -83,10 +85,11 @@ class MemoryStorage(
         """
         await self.get(key)
         await self.__schedule_expiry()
-        self.storage[key] += amount
+        async with self.locks[key]:
+            self.storage[key] += amount
 
-        if elastic_expiry or self.storage[key] == amount:
-            self.expirations[key] = time.time() + expiry
+            if elastic_expiry or self.storage[key] == amount:
+                self.expirations[key] = time.time() + expiry
 
         return self.storage.get(key, amount)
 
@@ -98,7 +101,8 @@ class MemoryStorage(
         """
         await self.get(key)
         await self.__schedule_expiry()
-        self.storage[key] = max(self.storage[key] - amount, 0)
+        async with self.locks[key]:
+            self.storage[key] = max(self.storage[key] - amount, 0)
 
         return self.storage.get(key, amount)
 
@@ -109,6 +113,7 @@ class MemoryStorage(
         if self.expirations.get(key, 0) <= time.time():
             self.storage.pop(key, None)
             self.expirations.pop(key, None)
+            self.locks.pop(key, None)
 
         return self.storage.get(key, 0)
 
@@ -119,6 +124,7 @@ class MemoryStorage(
         self.storage.pop(key, None)
         self.expirations.pop(key, None)
         self.events.pop(key, None)
+        self.locks.pop(key, None)
 
     async def acquire_entry(
         self, key: str, limit: int, expiry: int, amount: int = 1
@@ -274,5 +280,6 @@ class MemoryStorage(
         self.storage.clear()
         self.expirations.clear()
         self.events.clear()
+        self.locks.clear()
 
         return num_items