… CONTRIBUTIONS.rst
… doc/source/strategies.rst
… limits/aio/strategies.py
… limits/resources/redis/lua_scripts/sliding_window.lua
… limits/storage/base.py
… limits/storage/memcached.py
… limits/storage/memory.py
… limits/storage/redis.py
… limits/strategies.py
… limits/typing.py
… tests/aio/test_strategy.py
… tests/benchmarks/test_storage.py
… tests/integration/test_concurrency.py
… tests/test_strategy.py
… tests/utils.py