Following discussion https://github.com/alisaifee/limits/discussions/245

Sliding window counter implementation.
Supported backends: in-memory, redis, memcached.

TODO:
- [x] Floor the weighted counter to avoid premature rate limit. update documentation.
- [x] Memcached: implementation is sensitive to clock drift between limiter workers (time sync is mandatory)
- [ ] How to emulate slow IO (get, incr...) for any storage in tests?
- [ ] Memcached: why setting noreply=True leads to worse IOPS?
- [x] Memcached: remove expiry counter creation (not needed) to increase performance
- [x] In-memory (async): concurrency problem, can allow more requests than expected -> timestamp-based keys fix this issue
- [x] in-memory (sync): check if the window shift is ok against high concurrency / multi threaded environment. Otherwise use the timestamp based keys -> timestamp-based keys like the async version
- [x] Add tests: reset time memcached/other, weighted counter, concurrency?
- [x] compute reset time (different if memcached or other)
- [x] Remove test application, remove comments, clean code
- [x] Squash in one commit