@alisaifee 2025-01-10
Small tip: to fix linting issues locally before pushing you can run:
```
make lint-fix
```
and to just check:
```
make lint
```

(You'd ofcourse need all the development dependencies by running `pip install -r requirements/dev.txt`)