@Merinorus 2025-01-10
Thanks for the tip! I was doing the Ruff commands one by one, sometimes missing one.

We'll have to double-check the Memcached and Redis algorithms if they are valid on paper.
The performance on Redis and in memory seems on par with the moving window algorithm. The Memcached is a bit slower but, if it works, I will leave it as it is for now.

I have to check also the async in-memory storage concurrency, sometimes it allows one more hit than expected.
```
=================================== FAILURES ===================================
____________ TestConcurrency.test_sliding_window_counter[in-memory] ____________
tests/integration/test_concurrency.py:67: in test_sliding_window_counter
    assert len(hits) == 5
E   assert 6 == 5
E    +  where 6 = len([None, None, None, None, None, None])
=============================== warnings summary ===============================
```

What do you think about the current draft?