@alisaifee 2025-01-10
I haven't fully thought (or tested) this through, but just throwing it out there in case it makes sense.

For the current / previous key - could we use a different scheme where:
```py
now = int(time.time())
current_key = limit.key_for(*[*identifiers, str(int(now / item.get_expiry()))])
prev_key = limit.key_for(*[*identifiers, str(int((now - item.get_expiry()) / item.get_expiry()))])
```

And then use `item.get_expiry() * 2` for each key.

This way you wouldn't have to migrate the count from current to previous