@alisaifee 2025-01-10
Perhaps we could also push the onus of previous/current key into the storage so that the storage splits the key into two. This could have the additional benefit of only adding the `{` `}` brackets in the redis cluster implementation where it makes sense.