@Merinorus 2025-01-11
> I haven't fully thought (or tested) this through, but just throwing it out there in case it makes sense.
> 
> For the current / previous key - could we use a different scheme where:
> 
> ```python
> now = int(time.time())
> current_key = limit.key_for(*[*identifiers, str(int(now / item.get_expiry()))])
> prev_key = limit.key_for(*[*identifiers, str(int((now - item.get_expiry()) / item.get_expiry()))])
> ```
> 
> And then use `item.get_expiry() * 2` for each key.
> 
> This way you wouldn't have to migrate the count from current to previous

Yes indeed, it makes sense. I thought about this kind of solution for memcached.

Would you use this scheme on redis as well? What I like about the current solution is that it depends on the redis clock only, it's not mixed with the rate limit worker's clock. I'm not sure what is the convention, but I would be more confident using this clock, assuming redis cluster nodes have mechanisms to keep the clock in sync (and there should be no problem with a single redis instance, even if it's not synced).