@alisaifee 2025-01-11
> Would you use this scheme on redis as well? What I like about the current solution is that it depends on the redis clock only, it's not mixed with the rate limit worker's clock. 

If we agree on moving the splitting of the key into the storage implementation (i.e. the strategy doesn't know about which clock is being used) then we can perhaps fine tune this depending on the capabilities of the storage.