@Merinorus 2025-01-13
Hi,

So I moved the splitting logic into the storage implementation.
For the Memcached implementation, I did the following:

1. previous and current keys are computed according to the current time and the expiry time,
2. key shifting and lock are removed _(no performance gain, but not having to lock anything is comforting)_,
3. rate limiter hit attempts now need three Memcached variables instead of four _(almost no performance gain and complicated code, I intend to remove this part)_.

The limiter reset time is now more predictable, although this is no big deal.
However, the performance gain is negligible: the lock and time shift logic happened once per sampling period.

I did something else: for Memcached counter incrementation, a check was made at first with "add". I removed this check by doing counter incrementation directly. If I have a return value, the key already exists and I return the value. Otherwise, it doesn't exist yet and I re-run the exact counter logic written before. This makes the first incrementation slower, but all the subsequent writes are much faster.

With Memcached, the rate limiter median response time dropped from 1081 µs to 677 µs.
With the Memcached cluster, it dropped from 1274 µs to 1081 µs.

I think we can optimize the Memcached cluster response by having the current key and the previous key on the same node with some hashing logic.