@alisaifee 2025-01-14
> Let me know! I'll rework the async part when we agree on this one.

This looks great, in fact your changes to memcached is definitely a "more" correct approach (I couldn't reason why I'd chosen add first, but to be fair that was more than a decade ago 😁)

I'm happy to proceed with this approach. Once you're mostly comfortable with the memory, memcached & redis implementations for both sync & async we could consider moving this to master and I can take a crack at the mongodb implementation (I'm not sure anyone actually uses etcd so might not be worth the effort).