@Merinorus 2025-01-14
Thank you!
For the incrementation rework, while it's faster in case of DDoS attacks, it might be slower in normal use: when the user hits the limiter for the first time, it will need three memcached operations instead of two. Hence, the first implementation, maybe... But I guess it's better to favor IOPS when it's under pressure (DDoS, etc).

Sorry if it's not clear to me: is the current synchronous implementation of Redis and in-memory limiter okay for you? Or should I replicate the new Memcached version with the time-generated keys on these backbends as well?