@alisaifee 2025-01-14
> Sorry if it's not clear to me: is the current synchronous implementation of Redis and in-memory limiter okay for you? Or should I replicate the new Memcached version with the time-generated keys on these backbends as well?

My apologies for not clarifying, I think the current versions for in-memory & redis are fine (and in fact, I will probably use a similar approach for mongodb) since the atomicity of the operations (and the current/previous swap) can be managed on the server side.