@alisaifee 2025-01-20
Just FYI: this is the WIP for mongodb: https://github.com/alisaifee/limits/commit/f0534392af0667fe6f653a4ae041ab6e9eef19bc